Refactor Express server code
