+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "design-mode": "npm run build-css --watch & nodemon src/dev-server.js",
-    "start": "nodemon src/app.js",
-    "debug": "nodemon --inspect src/app.js",
+    "start": "nodemon src/server.js",
+    "debug": "nodemon --inspect src/server.js",
     "multi-tenant": "npm run build-css --watch & nodemon src/multi-tenant.js",
     "build-css": "node watcher.js",
     "build": "pkg ."
@@ -83,7 +83,7 @@
     "excludeModules": ["basic-ftp"]
   },
   "bin": {
-    "app": "src/app.js"
+    "app": "src/server.js"
   },
   "pkg": {
     "scripts": [
